+++ tests/test_common_errors.py
@@ -170,11 +170,11 @@ def test_reactant_uptake_gammas(Reactor, build_reactor):
     obj.reactant_uptake(hypothetical_gamma=[1e-7, 1e-8], disp=False)
 
     with pytest.raises(
-        TypeError, match=r"Gamma input must be int, float, or Array-like of int"
+        TypeError, match=r"Gamma input must be float or Array-like"
     ):
         obj.reactant_uptake(hypothetical_gamma="test", disp=False)
     with pytest.raises(
-        TypeError, match=r"Gamma input must be int, float, or Array-like of int"
+        TypeError, match=r"Gamma input must be float or Array-like"
     ):
         obj.reactant_uptake(hypothetical_gamma=["test", "test2"])