+++ docs/src/changelog.rst
@@ -1,8 +1,9 @@
 Changelog
 =========
 
-1.3.2
+1.4.0
 -------------------
+- Removed functionality for calculating the diffusion corrected uptake coefficient
 - Added support for changing individual parameters of the `boat_reactor` and `coated_wall_reactor` modules after initialization by automatically re-running the relevant calculations when a parameter is updated.
 - Added exposure time return for kinetics fitting functions in `boat_reactor` and `coated_wall_reactor` modules.
 - Added Lennard-Jones parameters for ClONO2, N2O5, O3, and NO2 to `diffusion_coef` module in order to support diffusion coefficient calculations for these species.
@@ -11,6 +12,7 @@ Changelog
 - Updated minimum carrier flow rate calculation to account for the presence of an insert
 - Added calculation of concentration inside the insert and used this concentration to calculate the fraction of unreacted surface sites after a given exposure time in the `coated_wall_reactor` module.
 - Fixed issue where a manually inputted diffusion coefficient would be overwritten if any attributes were updated.
+- Removed wall loss from `coated_wall_reactor` and moved it to optional for the `boat_reactor` module.
 - Miscellaneous improvements and bug fixes.
 
 1.3.1